pyproject.toml
[project]
name = "bigdata_mongo_taxi"
version = "0.1.0"
description = "Big Data NY Taxi + MongoDB Project"
requires-python = ">=3.10"
dependencies = [
    "polars-lts-cpu>=1.33.1",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pymongo>=3.13.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]

[tool.uv]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=9.0.1",
]
